services/prediction/requirements.txt
pymysql
pandas
scikit-learn
joblib
python-dotenv
flask
openai